requirements.txt
praw==7.7.1
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.2
nltk==3.8.1
yfinance==0.2.28
python-dotenv==1.0.0
flask==2.3.3
werkzeug==2.3.7
matplotlib==3.8.4
seaborn==0.13.2
textblob==0.18.0.post0 # Added textblob
cachetools==5.5.2 # Added cachetools